Implement changes needed to support unit testing of the loan input code.
First, the MultiScenarioInputValue needed the option to allocate managed objects using the in memory data context used by the test cases. It had previously been hard-coded to use the same SQLite database as the main application.

Another change is to pass as a parameter the scenario used by the loan classes. Previously, the loan classes were retrieving the scenario from "current input scenario" used in the production database. Passing in the parameter not only enables testing, but also will support running the simulator across multiple scenarios.