Fix a defect with editing of date sensitive values (for amounts, interest rates, growth rates), whereby editing the fixed value for a different scenario was causing an assertion failure in the table view code. There's a detailed comment in DateSensitiveValueFormInfoCreator describing the underlying problem and solution.
Along with this change, some refactoring was done for code that was "getting in the way" of a solution (but didn't end up being the part of the root cause). In particular, the MultiScenarioInputValueFieldInfo class was simplified, assuming it is passed an existing MultiScenarioInputValue, rather than passing in the parent object and the property key for the MultiScenarioInputValueFieldInfo's property. Since all classes which use MultiScenarioInputValueFieldInfo require the property to be initialized, this simplified code is possible.

The NumberFieldCell as also refactored to pass in pointers to the validator, field info, and formatter objects, rather than a pointer back to the NumberFieldEditInfo. A follow-on enhancement is suggested to bundle up these 3 properties and pass them as an "info" object.