Fix several small defects with the UI code for viewing and editing date sensitive value changes.
First, the way field edits were fixed to allow for other fields which are required, but have not yet been initialized. Instead of throwing an exception if the save fails, a "soft save" is performed which tries to save but doesn't generate an exception otherwise.

Secondly, there was a defect with the number fields, whereby if a value was already initialized, but then the field is cleared out, then the Save button is pressed,  a system generated exception is thrown. The fix was to propagate the disabling of fields down to the NumberFieldCell. When the NumberFieldCell is disabled, it will allow the cell to give itself up as first responder and gracefully exit.

Finally, there was a defect in the sorting of date sensitive value changes, since the start date had been migrated to a SimDate type which didn't support the compare selector. The fix was to define an overridden compare selector for the SimDate.