Refactored the SelectableObjectTableEditViewController class to maintain the selection information in the same way as the multiple selection class. This resulted in a simplified implementation, because the logic to maintain the selection (with deletes, indices moving, etc.) is greatly simplified.
Along with this implementation, a "start in edit" mode and default selection was added; this improves the ease of use for setting initial values.  Before, the user had to (1) Choose which value to edit, (2) Go in to edit mode, (2) Change the value (3) Leave edit mode, (4) Select the value just edited, (5) navigate back from the form. Now, the steps are just (2) and (5).