Fix an issue where a working balance was returning a positive balance before the balance start date.
The WorkingBalance class had a currentBalance property which was being read directly and incorporated in the results.
The problem with this is that the current balance is 0.0 before the balanceStartDate. The resolution was to remove the
externally visible currentBalance property and require clients of the class to get a current balance as of a given date.

This issue was surfacing in net worth calculations with loans that hadn't yet originated. A new unit test was added to
test this behavior, and existing ones were updated to call the new currentBalanceForDate selector.