Incorporate CorePlot and ResultraLib library projects as sub-projects of the main app project, rather than using Workspace.
This allows the dependencies to specified directly and explicitly in the main project, and stored under version control.
Dependencies for linking are also setup against the sub-projects. While I haven't don't an exhaustive comparison
of workspaces vs. sub-projects, this setup was found to work more smoothly on another App project.