Support the scenario where an asset value depreciates.
Previously an asset sold for a loss would trigger an assertion failure. The code was enhanced to support the asset loss.
This included handling a loss when a the asset is listed as something which is taxable. New unit tests have also been
added to cover the asset gain or loss scenario.