Only show "Amount Borrowed Growth Rate" when loan originates in the future.
The form for editing loan properties was previously showing this field, even if the loan originated in the past.
The main loan details form, the "what if" form for rates, and the form for editing scenario values has been
enhanced to only show this property if the loan originates in the future.

A context-sensitive help page has also been added to explain the loan origination parameters.