Don't prompt for a loan's starting balance when/if the loan originates in the future.
A loan's starting balance is not applicable when a loan originates in the future. After the loan's origination date is defined,
the starting balance will be shown if the loan originates in the past.

There is a special case where the user changes the starting
balance from the future to the past, leaving the starting balance undefined. In this case, the simulator will calculate a
starting balance based upon all the payments preceding the simulation start date.