Enhance the view for creating inputs to have an icon and subtitle for each input type.
This is important, since creating an input is the first thing a user will do, and gives this part of the UI more
polish. This also allows a subtitle giving a cue regarding the purpose of that input, and allowed
removal of a dedicated class for creating new inputs, in favor of using the standard form-based
table view classes.