Use periodic interest for loan amortization, rather than daily interest.
This is consistent with "by hand" loan amortizations. Using the daily interest calculations, there were minor discrepancies in the way interest was being calculated. With these change, the tests could also be cross-checked with a spreadsheet that uses standard loan amortization formulas. This spreadsheet has been checked in along with updated unit tests.