Prorate interest when the loan has an early payoff.
The principal change is to the PeriodicInterestBearingWorkingBalance class, which has been extended to include logic for track the last payment date for purposes of calculating the prorated interest.

This change also includes refactoring of the code to calculate loan interest to put the math into a common helper function. Other classes which perform loan interest type calculations all reference this helper code now.