Implement support for variable loan interest rates.
This involves a fair amount of refactoring in the loan simulation classes, principally PeriodicInterestBearingWorkingBalance. PeriodicInterestBearingWorkingBalance was enhanced with methods for decrementing the different types of payments, including interest only payments, regular periodic payments, skipped payments and the first non-deferred payment. For loans originating in the past, this refactoring also simplifies the logic to initialize the starting loan balance, which is done in LoanSimInfo. All the existing unit tests have been updated to pass with this new code, but new ones  have not yet been added for adjustable interest rates.