Don't require a current asset value if the purchase date is in the future.
Similar to what is already supported for loans, the current asset value is not even referenced if the asset purchase date is in the future. To eliminate confusion, it is hidden from view until the purchase date is set. If the purchase date is in the future, it is not shown. If it is set in the past, the field is shown.