Add explicit tie-break priorities for each of the event types, ensuring that if events happen on the same date, there is a sensible event order.
For, if a loan is originated on the same day as an asset purchase, the assumption is the money will be received before the money
is spent purchasing an asset. Similarly, if some income is received on the same day a transfer is made to transfer the income,
the income will always have a higher priority.

The SimEventList class has been updated to support event tie-break priorities based upon a normalized/rounded date. In other words,
if 2 event dates are on the same day, but one has a time later than the other, they will both be considered equal for tie-breaking purposes.