Customize the X Axis "unit label" for the different results.
Previously, the charts and tables were always showing "End of Year" as the label for the yearly value. However,
some of the values were a sum total for the year, so the result is not "as of the end of year". The plot data generators
for the different results type have been customized with different unit labels accordingly.