Implement some initial refinements to improve the performance of forecast generation.
First, the default length of the forecast has been changed to 25 years instead of 50. This is in consideration that the app is no longer strictly for retirement simulation. Secondly, after profiling the app, a major performance bottleneck was found in the 'processDigest' method, whereby date calculation was taking too long. Finally, for budgets which don't have tax inputs, the number of passes of digest calculations can be reduced from 3 to 1, which will improve forecast generation by 3 to 1 in those cases.