For performance improvement, enhance digest processing to only advance the account balances at the end of the year, or on days with digest entries.
Previously, the account balances were advanced every single day, which was taking up a lot of time in the forecast generation. This was considered necessary, since the taxes may include account interest. However, a reasonable trade-off is to ensure that account interest is calculated at the end of the year (and on days with other digest entries). For a reasonably comprehensive test case, this resulted in a 3-4x performance improvement (went from 2500ms to 775ms in the profiler).