Update the project file so that the builds can happen in a repeatable & standalone fashion w.r.t. to Resultra and 3rd party libraries.
This involved updating the header and library paths so they reference CorePlot in its "factory default" location of "core-plot" git clone. This also included updating some project references in the main FutureBudget project, so it would link with the CorePlot library and ResultraLibResources target. This should ensure with a clean checkout, FutureBudget can be built from scratch.